Refs #8523. Moved concurrent_queue.h
